requirements.txt
coverage==4.0.2
coveralls==1.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.4
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-envie==0.0.9
django-filter==0.13.0
django-nose==1.4.2
django-postgrespool==0.3.0
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.2.5
docopt==0.6.2
gunicorn==19.1.1
httpie==0.9.2
nose==1.3.7
psycopg2==2.5.1
Pygments==2.0.2
pytz==2016.1
PyYAML==3.11
requests==2.8.1
SQLAlchemy==1.0.9
sqlparse==0.1.18
static==0.4
static3==0.6.1
